refactor: UserContext, Converter 패키지 이동
