+++ src/main/java/woowacrew/article/controller/ArticleController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import woowacrew.article.domain.ArticleResponseDto;
 import woowacrew.article.service.ArticleService;
-import woowacrew.user.domain.UserContext;
+import woowacrew.user.dto.UserContext;
 import woowacrew.utils.annotation.AuthenticationUser;
 
 @Controller
