+++ src/main/java/woowacrew/user/controller/AdminApiController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import woowacrew.user.dto.UserApproveDto;
-import woowacrew.user.domain.UserContext;
+import woowacrew.user.dto.UserContext;
 import woowacrew.user.dto.UserResponseDto;
 import woowacrew.user.service.UserService;
