+++ src/main/java/woowacrew/user/controller/UserController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.RedirectView;
 import woowacrew.security.SecurityContextSupport;
-import woowacrew.user.domain.UserContext;
+import woowacrew.user.dto.UserContext;
 import woowacrew.user.dto.UserResponseDto;
 import woowacrew.user.dto.UserUpdateDto;
 import woowacrew.user.service.UserService;
