+++ src/main/java/woowacrew/user/service/UserService.java
@@ -3,8 +3,10 @@ package woowacrew.user.service;
 import org.springframework.stereotype.Service;
 import woowacrew.user.domain.*;
 import woowacrew.user.dto.UserApproveDto;
+import woowacrew.user.dto.UserContext;
 import woowacrew.user.dto.UserResponseDto;
 import woowacrew.user.dto.UserUpdateDto;
+import woowacrew.user.utils.UserConverter;
 
 import java.time.LocalDate;
 import java.util.List;
